Updated dryrun script * root level certificates

+++ bin/dryrun-app.sh
@@ -3,6 +3,8 @@ set -e
 
 full_path=${1}
 deployment=${2:-dev}
+keep_old=${3:-no}
+keep_old=$(echo "${keep_old:0:1}" | tr '[:upper:]' '[:lower:]')
 app=$(basename $full_path)
 location=$(dirname $full_path)
 tmp_path=".tmp/$location/$deployment-$app.yaml"
@@ -32,9 +34,16 @@ popd &> /dev/null
 echo $chart_checksum > $chart_checksum_path
 
 if [ -f "$tmp_path" ]; then
-    echo "  > found existing file"
-    echo "  > renaming old file to "$tmp_path".old"
-    mv $tmp_path $tmp_path".old"
+    echo "Found existing file"
+    if [[ -f "$tmp_path".old && "${keep_old:0:1}" == "y"  ]]; then
+        num_files=$(ls -1q "${tmp_path}.old"* 2> /dev/null | wc -l | tr -d '[:space:]')
+        echo " -> Renaming old file to "$tmp_path".old.${num_files}"
+        mv $tmp_path $tmp_path".old.${num_files}"
+    else
+        rm $tmp_path".old"*
+        echo " -> Renaming old file to "$tmp_path".old"
+        mv $tmp_path $tmp_path".old"
+    fi
 fi
 
 echo "  > running..."
